[#510] Update cluster module references in code
